+++ src/negotiation/execution_params.ts
@@ -1,7 +1,6 @@
 import moment = require("moment");
 
-// Information that cannot be deduced from the OrderParams
-// and is not up for negotiation
+// Information that cannot be deduced from the order and is not up for negotiation
 // Matches exactly cnd http api
 export interface ExecutionParams {
   peer: {