+++ src/negotiation/taker/maker_client.ts
@@ -1,22 +1,18 @@
 import axios from "axios";
-import { ExecutionParams } from "./execution_params";
-import { OrderParams } from "./order";
+import { ExecutionParams } from "../execution_params";
+import { Order } from "../order";
 
 export class MakerClient {
   constructor(private readonly makerUrl: string) {}
 
-  public async getOrderByTradingPair(
-    tradingPair: string
-  ): Promise<OrderParams> {
+  public async getOrderByTradingPair(tradingPair: string): Promise<Order> {
     const response = await axios.get(`${this.makerUrl}orders/${tradingPair}`);
     return response.data;
   }
 
-  public async getExecutionParams(
-    orderParams: OrderParams
-  ): Promise<ExecutionParams> {
+  public async getExecutionParams(order: Order): Promise<ExecutionParams> {
     const response = await axios.get(
-      `${this.makerUrl}orders/${orderParams.id}/executionParams`
+      `${this.makerUrl}orders/${order.id}/executionParams`
     );
     return response.data;
   }