+++ src/negotiation/taker/negotiator.spec.ts
@@ -1,37 +1,37 @@
-import { MockBitcoinWallet } from "../__mocks__/bitcoin_wallet";
-import { Cnd } from "../cnd";
-import { ComitClient } from "../comit_client";
-import { EthereumWallet } from "../ethereum_wallet";
+import { MockBitcoinWallet } from "../../__mocks__/bitcoin_wallet";
+import { Cnd } from "../../cnd";
+import { ComitClient } from "../../comit_client";
+import { EthereumWallet } from "../../ethereum_wallet";
 import {
   // @ts-ignore: tslint does not know that this is actually ./__mocks__/maker_client
   mockGetExecutionParams,
   // @ts-ignore: tslint does not know that this is actually ./__mocks__/maker_client
   mockGetOrderByTradingPair
 } from "./maker_client";
-import { TakerCriteria } from "./order";
-import { TakerNegotiator } from "./taker_negotiator";
+import { Negotiator } from "./negotiator";
+import { MatchingCriteria } from "./order";
 
 jest.mock("./maker_client");
-jest.mock("../ethereum_wallet");
-jest.mock("../cnd");
+jest.mock("../../ethereum_wallet");
+jest.mock("../../cnd");
 
 const defaultCnd = new Cnd("");
-const defaultComitClient = new ComitClient(defaultCnd);
-defaultComitClient.bitcoinWallet = new MockBitcoinWallet();
-defaultComitClient.ethereumWallet = new EthereumWallet("");
+const defaultComitClient = new ComitClient(defaultCnd)
+  .withBitcoinWallet(new MockBitcoinWallet())
+  .withEthereumWallet(new EthereumWallet(""));
 
-describe("Taker Negotiator", () => {
+describe("taker.Negotiator", () => {
   beforeEach(() => {
     mockGetOrderByTradingPair.mockClear();
     mockGetExecutionParams.mockClear();
   });
 
   it("Search an order and takes it", async () => {
-    const takerNegotiator = new TakerNegotiator(defaultComitClient, "");
+    const takerNegotiator = new Negotiator(defaultComitClient, "");
 
     // Define criteria for the order such as the assets and ledgers
     // But also min/max amounts we are looking to trade
-    const takerCriteria: TakerCriteria = {
+    const matchingCriteria: MatchingCriteria = {
       buy: {
         ledger: "bitcoin",
         asset: "bitcoin",
@@ -44,7 +44,7 @@ describe("Taker Negotiator", () => {
     };
 
     // Ask the maker if there are any orders that matches the criteria
-    const order = await takerNegotiator.getOrder(takerCriteria);
+    const order = await takerNegotiator.getOrder(matchingCriteria);
 
     // We expect the returned order to be valid, if not
     // this maker may be broken or malicious
@@ -62,9 +62,9 @@ describe("Taker Negotiator", () => {
   });
 
   it("Returns an order with the correct trading pair, given buy/sell input", async () => {
-    const takerNegotiator = new TakerNegotiator(defaultComitClient, "");
+    const takerNegotiator = new Negotiator(defaultComitClient, "");
 
-    const takerCriteria: TakerCriteria = {
+    const matchingCriteria: MatchingCriteria = {
       buy: {
         ledger: "bitcoin",
         asset: "bitcoin",
@@ -76,12 +76,12 @@ describe("Taker Negotiator", () => {
       }
     };
 
-    const order = await takerNegotiator.getOrder(takerCriteria);
+    const order = await takerNegotiator.getOrder(matchingCriteria);
 
-    expect(order.orderParams).toHaveProperty("bid.ledger", "bitcoin");
-    expect(order.orderParams).toHaveProperty("bid.asset", "bitcoin");
-    expect(order.orderParams).toHaveProperty("ask.ledger", "ethereum");
-    expect(order.orderParams).toHaveProperty("ask.asset", "ether");
+    expect(order.rawOrder).toHaveProperty("bid.ledger", "bitcoin");
+    expect(order.rawOrder).toHaveProperty("bid.asset", "bitcoin");
+    expect(order.rawOrder).toHaveProperty("ask.ledger", "ethereum");
+    expect(order.rawOrder).toHaveProperty("ask.asset", "ether");
 
     expect(mockGetOrderByTradingPair).toHaveBeenCalledTimes(1);
     expect(mockGetExecutionParams).toHaveBeenCalledTimes(0);